Resplit Assets to hosts and oss, add relevant mixins
