+++ gvm/protocols/gmpv208/gmpv208.py
@@ -601,67 +601,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def get_tls_certificates(
-        self,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        include_certificate_data: Optional[bool] = None,
-        details: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of TLS certificates
-
-        Arguments:
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            include_certificate_data: Whether to include the certificate data in
-                the response
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-
-        cmd = XmlCommand("get_tls_certificates")
-
-        add_filter(cmd, filter, filter_id)
-
-        if details is not None:
-            cmd.set_attribute("details", to_bool(details))
-
-        if include_certificate_data is not None:
-            cmd.set_attribute(
-                "include_certificate_data", to_bool(include_certificate_data)
-            )
-
-        return self._send_xml_command(cmd)
-
-    def get_tls_certificate(self, tls_certificate_id: str) -> Any:
-        """Request a single TLS certificate
-
-        Arguments:
-            tls_certificate_id: UUID of an existing TLS certificate
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        cmd = XmlCommand("get_tls_certificates")
-
-        if not tls_certificate_id:
-            raise RequiredArgument(
-                function=self.get_tls_certificate.__name__,
-                argument='tls_certificate_id',
-            )
-
-        cmd.set_attribute("tls_certificate_id", tls_certificate_id)
-
-        # for single tls certificate always request cert data
-        cmd.set_attribute("include_certificate_data", "1")
-
-        # for single entity always request all details
-        cmd.set_attribute("details", "1")
-
-        return self._send_xml_command(cmd)
-
     def modify_audit(
         self,
         audit_id: str,
@@ -992,45 +931,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def modify_tls_certificate(
-        self,
-        tls_certificate_id: str,
-        *,
-        name: Optional[str] = None,
-        comment: Optional[str] = None,
-        trust: Optional[bool] = None,
-    ) -> Any:
-        """Modifies an existing TLS certificate.
-
-        Arguments:
-            tls_certificate_id: UUID of the TLS certificate to be modified.
-            name: Name of the TLS certificate, defaulting to the MD5 fingerprint
-            comment: Comment for the TLS certificate.
-            trust: Whether the certificate is trusted.
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not tls_certificate_id:
-            raise RequiredArgument(
-                function=self.modify_tls_certificate.__name__,
-                argument='tls_certificate_id',
-            )
-
-        cmd = XmlCommand("modify_tls_certificate")
-        cmd.set_attribute("tls_certificate_id", str(tls_certificate_id))
-
-        if comment:
-            cmd.add_element("comment", comment)
-
-        if name:
-            cmd.add_element("name", name)
-
-        if trust:
-            cmd.add_element("trust", to_bool(trust))
-
-        return self._send_xml_command(cmd)
-
     def clone_ticket(self, ticket_id: str) -> Any:
         """Clone an existing ticket
 
@@ -1048,27 +948,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd = XmlCommand("create_ticket")
 
         _copy = cmd.add_element("copy", ticket_id)
-
-        return self._send_xml_command(cmd)
-
-    def clone_tls_certificate(self, tls_certificate_id: str) -> Any:
-        """Modifies an existing TLS certificate.
-
-        Arguments:
-            tls_certificate_id: The UUID of an existing TLS certificate
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not tls_certificate_id:
-            raise RequiredArgument(
-                function=self.clone_tls_certificate.__name__,
-                argument='tls_certificate_id',
-            )
-
-        cmd = XmlCommand("create_tls_certificate")
-
-        cmd.add_element("copy", tls_certificate_id)
 
         return self._send_xml_command(cmd)
 
@@ -1337,23 +1216,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd = XmlCommand("delete_config")
         cmd.set_attribute("config_id", policy_id)
         cmd.set_attribute("ultimate", to_bool(ultimate))
-
-        return self._send_xml_command(cmd)
-
-    def delete_tls_certificate(self, tls_certificate_id: str) -> Any:
-        """Deletes an existing tls certificate
-
-        Arguments:
-            tls_certificate_id: UUID of the tls certificate to be deleted.
-        """
-        if not tls_certificate_id:
-            raise RequiredArgument(
-                function=self.delete_tls_certificate.__name__,
-                argument='tls_certificate_id',
-            )
-
-        cmd = XmlCommand("delete_tls_certificate")
-        cmd.set_attribute("tls_certificate_id", tls_certificate_id)
 
         return self._send_xml_command(cmd)
